Added Django REST framework stuff
Added some stuff to have DRF working with the template. An application
named "api" has been added to contain all the logic of the microservice.
It's been also added a serviceaccessible under /api/status/ (GET method)
to return the status of the service.

See README.md for more information about the decissions made.